refactor: split container_execution/mod.rs into submodules (#130)
Split the 1208-line mod.rs into four focused submodules:
- docker_args.rs: build_docker_args, container_name_for
- infra.rs: detect_compose_network, make_mcp_config, token helpers
- process.rs: StreamResult, run_container, run_iteration, stream helpers
- mod.rs: re-exports, ExecutionConfig, IterationOutcome, STREAM_DISPLAY_JQ

Public interface from capsule::container_execution::* unchanged.
All 178 tests pass, clippy clean.
